+++ App/PHTVInputMethod/Application/main.swift
@@ -1,7 +1,0 @@
-import Cocoa
-
-let application = NSApplication.shared
-let delegate = PHTVInputMethodApplicationDelegate()
-application.delegate = delegate
-
-_ = NSApplicationMain(CommandLine.argc, CommandLine.unsafeArgv)